Remove compat below API level 7.
