+++ src/com/ichi2/anki/CardEditor.java
@@ -835,9 +835,7 @@ public class CardEditor extends ActionBarActivity {
                     intent.putExtra(EXTRA_CONTENTS, getFieldsText());
                 }
                 startActivityForResult(intent, REQUEST_ADD);
-                if (AnkiDroidApp.SDK_VERSION > 4) {
-                    ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.LEFT);
-                }
+                ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.LEFT);
                 return true;
 
             case MENU_RESET:
@@ -943,17 +941,11 @@ public class CardEditor extends ActionBarActivity {
         finish();
         if (mCaller == CALLER_CARDEDITOR_INTENT_ADD || mCaller == CALLER_BIGWIDGET_EDIT
                 || mCaller == CALLER_BIGWIDGET_ADD) {
-            if (AnkiDroidApp.SDK_VERSION > 4) {
-                ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.FADE);
-            }
+            ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.FADE);
         } else if (mCaller == CALLER_INDICLASH) {
-            if (AnkiDroidApp.SDK_VERSION > 4) {
-                ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.NONE);
-            }
+            ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.NONE);
         } else {
-            if (AnkiDroidApp.SDK_VERSION > 4) {
-                ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.RIGHT);
-            }
+            ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.RIGHT);
         }
     }
 
@@ -1187,9 +1179,7 @@ public class CardEditor extends ActionBarActivity {
                 intent.putExtra(EXTRA_CONTENTS, map.get("fields"));
                 intent.putExtra(EXTRA_ID, map.get("id"));
                 startActivityForResult(intent, REQUEST_INTENT_ADD);
-                if (AnkiDroidApp.SDK_VERSION > 4) {
-                    ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.FADE);
-                }
+                ActivityTransitionAnimation.slide(CardEditor.this, ActivityTransitionAnimation.FADE);
                 mIntentInformationDialog.dismiss();
             }
         });
