+++ src/com/ichi2/anki/CramDeckOptions.java
@@ -357,7 +357,7 @@ public class CramDeckOptions extends PreferenceActivity implements OnSharedPrefe
         // This screen doesn't suffer from this bug really as it doesn't have nested
         // PreferenceScreens, but we have change the background for consistency with
         // DeckOptions background.
-        if (AnkiDroidApp.SDK_VERSION >= 7 && AnkiDroidApp.SDK_VERSION <= 10) {
+        if (AnkiDroidApp.SDK_VERSION <= 10) {
             Themes.applyTheme(this, Themes.THEME_ANDROID_DARK);
         }
         super.onCreate(icicle);
@@ -409,9 +409,7 @@ public class CramDeckOptions extends PreferenceActivity implements OnSharedPrefe
         if (keyCode == KeyEvent.KEYCODE_BACK && event.getRepeatCount() == 0) {
             Log.i(AnkiDroidApp.TAG, "DeckOptions - onBackPressed()");
             finish();
-            if (AnkiDroidApp.SDK_VERSION > 4) {
-                ActivityTransitionAnimation.slide(this, ActivityTransitionAnimation.FADE);
-            }
+            ActivityTransitionAnimation.slide(this, ActivityTransitionAnimation.FADE);
             return true;
         }
         return super.onKeyDown(keyCode, event);
