+++ src/com/ichi2/anki/Preferences.java
@@ -125,7 +125,7 @@ public class Preferences extends PreferenceActivity implements OnSharedPreferenc
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         // Workaround for bug 4611: http://code.google.com/p/android/issues/detail?id=4611
-        if (AnkiDroidApp.SDK_VERSION >= 7 && AnkiDroidApp.SDK_VERSION <= 10) {
+        if (AnkiDroidApp.SDK_VERSION <= 10) {
             Themes.applyTheme(this, Themes.THEME_ANDROID_DARK);
         }
         super.onCreate(savedInstanceState);
@@ -209,10 +209,6 @@ public class Preferences extends PreferenceActivity implements OnSharedPreferenc
             updateNumberRangePreference(key);
         }
 
-        if (AnkiDroidApp.SDK_VERSION <= 4) {
-            fadeScrollbars.setChecked(false);
-            fadeScrollbars.setEnabled(false);
-        }
     }
 
 
@@ -550,9 +546,7 @@ public class Preferences extends PreferenceActivity implements OnSharedPreferenc
 
     private void closePreferences() {
         finish();
-        if (AnkiDroidApp.SDK_VERSION > 4) {
-            ActivityTransitionAnimation.slide(this, ActivityTransitionAnimation.FADE);
-        }
+        ActivityTransitionAnimation.slide(this, ActivityTransitionAnimation.FADE);
         if (mCol != null) {
             mCol.save();
         }
