+++ src/com/ichi2/compat/CompatV7.java
@@ -2,18 +2,45 @@ package com.ichi2.compat;
 
 import android.annotation.TargetApi;
 import android.app.Activity;
+import android.database.sqlite.SQLiteDatabase;
 import android.graphics.drawable.ColorDrawable;
+import android.media.AudioManager;
+import android.speech.tts.TextToSpeech;
+import android.speech.tts.TextToSpeech.OnUtteranceCompletedListener;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarActivity;
 import android.text.SpannableStringBuilder;
 import android.text.style.CharacterStyle;
 import android.text.style.ForegroundColorSpan;
+import android.view.View;
+import android.webkit.WebView;
 
 import com.ichi2.anki.R;
+import com.ichi2.anki.ReadText;
 
 /** Implementation of {@link Compat} for SDK level 7 */
 @TargetApi(7)
-public class CompatV7 extends CompatV5 implements Compat {
+public class CompatV7 implements Compat {
+
+    @Override
+    public String normalizeUnicode(String txt) {
+        return txt;
+    }
+
+    @Override
+    public void setScrollbarFadingEnabled(WebView webview, boolean enable) {
+        webview.setScrollbarFadingEnabled(enable);
+    }
+
+    @Override
+    public void setOverScrollModeNever(View v) {
+    }
+
+    @Override
+    public void invalidateOptionsMenu(Activity activity) {
+        ActionBarActivity actionBarActivity = (ActionBarActivity) activity;
+        actionBarActivity.supportInvalidateOptionsMenu();
+    }
 
     @Override
     public void setActionBarBackground(Activity activity, int color) {
@@ -58,9 +85,28 @@ public class CompatV7 extends CompatV5 implements Compat {
     }
 
     @Override
-    public void invalidateOptionsMenu(Activity activity) {
-        ActionBarActivity actionBarActivity = (ActionBarActivity) activity;
-        actionBarActivity.supportInvalidateOptionsMenu();
+    public void setTtsOnUtteranceProgressListener(TextToSpeech tts) {
+        tts.setOnUtteranceCompletedListener(new OnUtteranceCompletedListener() {
+            @Override
+            public void onUtteranceCompleted(String utteranceId) {
+                if (ReadText.sTextQueue.size() > 0) {
+                    String[] text = ReadText.sTextQueue.remove(0);
+                    ReadText.speak(text[0], text[1]);
+                }
+            }
+        });
+    }
+
+    @Override
+    public void disableDatabaseWriteAheadLogging(SQLiteDatabase db) {
+    }
+
+    @Override
+    public void requestAudioFocus(AudioManager audioManager) {
+    }
+
+    @Override
+    public void abandonAudioFocus(AudioManager audioManager) {
     }
 
 }
